requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
Django==4.1.5
django-froala-editor==4.0.16
django-tinymce==3.5.0
html2text==2020.1.16
htmldocx==0.0.6
lxml==4.9.2
python-docx==0.8.11
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
